application/requirements.txt
aiosmtplib==2.0.1
alembic==1.10.4
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
databases==0.7.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.0.2
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
psycopg2-binary==2.9.5
pydantic==1.10.7
PyJWT==2.6.0
python-decouple==3.7
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.22.0
typing_extensions==4.5.0
uvicorn==0.20.0
